feature: creating product

+++ Samples/Playlists/cs/Models/Product.cs
@@ -18,7 +18,7 @@ namespace Models
         public virtual decimal? DiscountPer { get; set; }
         [Required]
         public string Name { get; set; }
-        public decimal RefillTime { get; set; }
+        public Int32 RefillTime { get; set; }
         public virtual decimal? SGSTPer { get; set; }
         public decimal Threshold { get; set; }
         public decimal TotalQuantity { get; set; }
